Rename src/__tests__/sector to cad (#212)
