+++ viewer/src/__tests__/models/cad/determineSectors.test.ts
