+++ viewer/src/__tests__/models/cad/determineSectorsQuality.test.ts
