+++ viewer/src/__tests__/models/cad/emptySector.ts
