+++ viewer/src/__tests__/models/cad/initializeSectorLoader.test.ts
